Fix sngive admin commands

+++ src/main/java/org/uwu_snek/shadownight/adminCommands/CMD_sngive.java
@@ -7,6 +7,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.uwu_snek.shadownight.items.ItemManager;
+import org.uwu_snek.shadownight.utils.spigot.ChatUtils;
 
 
 
@@ -14,22 +15,21 @@ import org.uwu_snek.shadownight.items.ItemManager;
 public final class CMD_sngive implements CommandExecutor {
     @Override
     public boolean onCommand(final @NotNull CommandSender sender, final @NotNull Command command, final @NotNull String label, final @NotNull String @NotNull [] args) {
-        long id;
-        int amount;
         if(args.length > 0) {
-            final Player player = (Player) sender;
-            try {
-                id = Long.parseLong(args[0]);
-                amount = args.length > 1 ? Integer.parseInt(args[1]) : 1;
+            final Player player = (Player)sender;
+            final String pattern = String.join(" ", args);
 
-                ItemStack item = ItemManager.getValueFromId(id).createDefaultItemStack();
-                item.setAmount(amount);
-                player.getInventory().setItem(player.getInventory().firstEmpty(), item);
+            for(ItemManager manager : ItemManager.values()) {
+                if(
+                    ChatUtils.stripColor(manager.getInstance().getDisplayName()).toLowerCase().strip().replaceAll("\\s+", " ")
+                    .matches(pattern.toLowerCase().strip().replaceAll("\\s+", " "))
+                ) {
+                    ItemStack item = manager.getInstance().createDefaultItemStack();
+                    player.getInventory().setItem(player.getInventory().firstEmpty(), item);
+                }
             }
-            catch(NumberFormatException e) {
-                player.sendMessage("§cInvalid ID or amount. Usage: /sngive <ID> <amount>");
-            }
+            return true;
         }
-        return true;
+        return false;
     }
 }
